Ease extension of the Converter through types and more hooks
This pulls in much more of srcomp's typing to clarify the types
being used here and extracts more hooks to aid both with typing
and extension of this base. I'd still like to get to a point
where much more of this is configurable rather than needing
fully spelling out, but this is still a useful step.